+++ modules/juce_core/native/juce_linux_Network.cpp
@@ -127,7 +127,7 @@ public:
         if (select (socketHandle + 1, &readbits, 0, 0, &tv) <= 0)
             return 0;   // (timeout)
 
-        const int bytesRead = jmax (0, (int) recv (socketHandle, buffer, bytesToRead, MSG_WAITALL));
+        const int bytesRead = jmax (0, (int) recv (socketHandle, buffer, (size_t) bytesToRead, MSG_WAITALL));
         if (bytesRead == 0)
             finished = true;
 
@@ -193,7 +193,7 @@ private:
         else if (timeOutMs < 0)
             timeOutTime = 0xffffffff;
         else
-            timeOutTime += timeOutMs;
+            timeOutTime += (uint32) timeOutMs;
 
         String hostName, hostPath;
         int hostPort;
@@ -391,12 +391,12 @@ private:
 
             const int numToSend = jmin (1024, (int) (requestHeader.getSize() - totalHeaderSent));
 
-            if (send (socketHandle, static_cast <const char*> (requestHeader.getData()) + totalHeaderSent, numToSend, 0) != numToSend)
+            if (send (socketHandle, static_cast <const char*> (requestHeader.getData()) + totalHeaderSent, (size_t) numToSend, 0) != numToSend)
                 return false;
 
-            totalHeaderSent += numToSend;
+            totalHeaderSent += (size_t) numToSend;
 
-            if (progressCallback != nullptr && ! progressCallback (progressCallbackContext, totalHeaderSent, requestHeader.getSize()))
+            if (progressCallback != nullptr && ! progressCallback (progressCallbackContext, (int) totalHeaderSent, (int) requestHeader.getSize()))
                 return false;
         }
